Make sure that install_ansible script works on RHEL systems

+++ vagrant/install_ansible.sh
@@ -7,10 +7,16 @@ if [ $INSTALLED == '0' ]; then
     echo "Ansible installed"
 else
     echo "Installing Ansible..."
-    apt-get install -y software-properties-common
-    apt-add-repository ppa:ansible/ansible
-    apt-get update
-    apt-get install -y --force-yes ansible
+    if [ -f /etc/lsb-release ]; then
+        apt-get install -y software-properties-common
+        apt-add-repository ppa:ansible/ansible
+        apt-get update
+        apt-get install -y --force-yes ansible
+    else if [ -f /etc/redhat-release ]; then
+        yum install -y epel-release
+        yum install -y ansible
+        fi
+    fi
 fi
 
 cp /vagrant/ansible.cfg /home/vagrant/ansible.cfg
